Fix job controller, locales and specs
