+++ .gitignore
@@ -53,3 +53,4 @@ build-iPhoneSimulator/
 .byebug_history
 .idea/
 log/
+.ruby-version
