+++ config/routes.rb
@@ -8,7 +8,7 @@ Rails.application.routes.draw do
     get '/about', to: 'static_pages#about'
     get '/terms', to: 'static_pages#terms'
     get '/contact_us', to: 'static_pages#contact_us'
-  end
 
-  resources :jobs
+    resources :jobs
+  end
 end
