+++ spec/rails_helper.rb
@@ -9,6 +9,7 @@ require 'rspec/rails'
 
 require 'support/factory_bot'
 require 'support/shoulda_matchers'
+require 'support/sorcery'
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
