+++ spec/support/sorcery.rb
@@ -1,0 +1,4 @@
+RSpec.configure do |config|
+  config.include Sorcery::TestHelpers::Rails::Controller, type: :controller
+  config.include Sorcery::TestHelpers::Rails::Integration, type: :feature
+end
